Updated core files with fixes + cleanup
I cleaned up the code a bit for things I thought might be useful in the future but which I now realize just take up space. I also fixed the positioning and font scaling, so it might actually work now on all screens! Hooray! Also re-did some of the html so it's more compact. And I'm planning on adding a button using a map to turn the color separation of the text on/off as this is the most resource intensive process.